src/requirements.txt
pandas==1.3.0
scikit-learn==0.24.2
os
json
psycopg2-binary==2.9.2
pandas==1.3.4
sentence-transformers==2.1.0

